[C] Remove manual casting as the error was caused by the AERON_HDR_TYPE_EXT declaration issue.

+++ aeron-client/src/test/c/aeron_controlled_image_fragment_assembler_test.cpp
@@ -218,7 +218,7 @@ TEST_F(ControlledImageFragmentAssemblerTest, shouldReassembleFromTwoFragments)
         -190,
         63456385384L,
         fragmentLength,
-        (uint8_t)(fragmentLength % 256));
+        fragmentLength % 256);
     EXPECT_EQ(AERON_ACTION_CONTINUE, handle_fragment(handler, fragmentLength));
     EXPECT_TRUE(isCalled);
 }
@@ -269,12 +269,12 @@ TEST_F(ControlledImageFragmentAssemblerTest, shouldReassembleFromThreeFragments)
     EXPECT_FALSE(isCalled);
 
     termOffset += MTU_LENGTH;
-    fillFrame(0xE, termOffset, fragmentLength, (uint8_t)(fragmentLength % 256));
+    fillFrame(0xE, termOffset, fragmentLength, fragmentLength % 256);
     EXPECT_EQ(AERON_ACTION_CONTINUE, handle_fragment(handler, fragmentLength));
     EXPECT_FALSE(isCalled);
 
     termOffset += MTU_LENGTH;
-    fillFrame(0x7F, termOffset, lastFragmentLength, (uint8_t)((fragmentLength * 2) % 256));
+    fillFrame(0x7F, termOffset, lastFragmentLength, (fragmentLength * 2) % 256);
     handle_fragment(handler, lastFragmentLength);
     EXPECT_TRUE(isCalled);
 }
@@ -310,7 +310,7 @@ TEST_F(ControlledImageFragmentAssemblerTest, shouldNotReassembleIfMissingBegin)
     EXPECT_FALSE(isCalled);
 
     termOffset += MTU_LENGTH;
-    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, (uint8_t)(fragmentLength % 256));
+    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, fragmentLength % 256);
     EXPECT_EQ(AERON_ACTION_CONTINUE, handle_fragment(handler, fragmentLength));
     EXPECT_FALSE(isCalled);
 }
@@ -336,7 +336,7 @@ TEST_F(ControlledImageFragmentAssemblerTest, shouldReassembleTwoMessagesFromFour
     EXPECT_FALSE(isCalled);
 
     termOffset += MTU_LENGTH;
-    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, (uint8_t)(fragmentLength % 256));
+    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, fragmentLength % 256);
     EXPECT_EQ(AERON_ACTION_CONTINUE, handle_fragment(handler, fragmentLength));
     EXPECT_TRUE(isCalled);
 
@@ -349,7 +349,7 @@ TEST_F(ControlledImageFragmentAssemblerTest, shouldReassembleTwoMessagesFromFour
     EXPECT_FALSE(isCalled);
 
     termOffset += MTU_LENGTH;
-    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, (uint8_t)(fragmentLength % 256));
+    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, fragmentLength % 256);
     EXPECT_EQ(AERON_ACTION_CONTINUE, handle_fragment(handler, fragmentLength));
     EXPECT_TRUE(isCalled);
 
@@ -379,7 +379,7 @@ TEST_F(ControlledImageFragmentAssemblerTest, shouldAbortReassembly)
     EXPECT_FALSE(isCalled);
 
     termOffset += MTU_LENGTH;
-    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, (uint8_t)(fragmentLength % 256));
+    fillFrame(AERON_DATA_HEADER_END_FLAG, termOffset, fragmentLength, fragmentLength % 256);
     EXPECT_EQ(AERON_ACTION_ABORT, handle_fragment(handler, fragmentLength));
     EXPECT_TRUE(isCalled);
